[FIX] Should only use statement line reference for matching
In finnish banking contexts it is not possible to generate bank
statements with references in random positions: they are mostly
required. As such it makes no sense to fall back to the label of the
statement line if the reference is missing: in the rare cases that the
reference is missing, finnish users do not want the system to do
automatic reconciliation, because it is an exception rather than a
rule.