[REF] Refactor reconciliation tests to not install CoA
Change the base unit tests to be more in line with the way Odoo's own
account module does testing, which is to skip tests if no chart of
accounts is found. The old way the setup worked tried to install
a random chart of accounts, which set the system into an infinite
loop on my system.

One of the test cases was also broken, and was corrected